// src/constants/enums.ts
/*
 * enums.ts
 * Copyright: Ouranos Studio 2019. All rights reserved.
 */

export enum Services {
	app = 'app',
	panel = 'panel',
}

export enum LANGUAGE_CODES {
	en = 'en',
	fa = 'fa',
}

export enum COUNTRIES {
	us = 'U.S'
}

export enum GENDER {
	male = 'Male',
	female = 'Female',
}

export enum ACTIVITY {
	sed = 'sed',
	light = 'light',
	mod = 'mod',
	high = 'high',
	extreme = 'extreme',
}

export enum GOALS {
	ml = 'ml',
	sl = 'sl',
	il = 'il',
	m = 'm',
	mg = 'mg',
	sg = 'sg',
	ig = 'ig',
}

export enum STATUS {
	active = 'ACTIVE',
	inactive = 'INACTIVE',
}

export enum WEIGHT_UNITS {
	kg = 'kg',
	pound = 'pound',
}

export enum HEIGHT_UNITS {
	cm = 'cm',
}

export enum WEEKDAYS {
	saturday = 'saturday',
	sunday = 'sunday',
	monday = 'monday',
	tuesday = 'tuesday',
	wednesday = 'wednesday',
	thursday = 'thursday',
	friday = 'friday',
}

export enum EVENT_TYPES {
	meal = 'meal',
	exercise = 'exercise',
}


export enum FOOD_TYPES {
	type1 = 'Type 1',
	type2 = 'Type 2',
	unknown = 'Unkown',
}

export enum TAG_TYPE {
	cuisine = 'cuisine',
	meal = 'meal',
	diet = 'diet',
	recipe = 'recipe',
	ingredient = 'ingredient',
	occasion = 'occasion',
}

export enum MEAL_ITEM_TYPES {
	recipe = 'recipe',
	food = 'food',
}